Add basic drag & drop node

+++ src/TreeView.TreeView.ContextMenuStrip.Event.cs
@@ -13,12 +13,41 @@
                         treeviewControl.existingItemToolStripMenuItem.Click += new System.EventHandler(Folder.AddExistingItem);
                         treeviewControl.newFolderToolStripMenuItem.Click += new System.EventHandler(Folder.AddNew);
 
+                        treeviewControl.renameToolStripMenuItem1.Click += new System.EventHandler(Rename);
                         treeviewControl.deleteToolStripMenuItem.Click += new System.EventHandler(Folder.Delete);
+                        
 
+                        treeviewControl.openFolderInFileExplorerToolStripMenuItem.Click += new System.EventHandler(Open);
 
-                        treeviewControl.openToolStripMenuItem.Click += new System.EventHandler(File.Open);
+                        treeviewControl.openToolStripMenuItem.Click += new System.EventHandler(Open);
 
+                        treeviewControl.renameToolStripMenuItem1.Click += new System.EventHandler(Rename);
                         treeviewControl.deleteToolStripMenuItem1.Click += new System.EventHandler(File.Delete);
+                    }
+
+                    static public void Open(object sender, System.EventArgs e)
+                    {
+                        try
+                        {
+                            System.Diagnostics.Process.Start(treeviewControl.treeView.SelectedNode.Name);
+                        }
+                        catch (System.Exception exception)
+                        {
+                            System.Windows.Forms.MessageBox.Show(exception.Message, "List");
+                        }
+                    }
+
+                    static public void Rename(object sender, System.EventArgs e)
+                    {
+                        try
+                        {
+                            treeviewControl.treeView.LabelEdit = true;
+                            treeviewControl.treeView.SelectedNode.BeginEdit();
+                        }
+                        catch (System.Exception exception)
+                        {
+                            System.Windows.Forms.MessageBox.Show(exception.Message, "List");
+                        }
                     }
 
                     static private class Folder
@@ -36,11 +65,11 @@
                                 {
                                     try
                                     {
-                                        System.IO.File.Copy(path, treeviewControl.treeView.SelectedNode.Name + @"\" + System.IO.Path.GetFileName(path), false);
+                                        Microsoft.VisualBasic.FileIO.FileSystem.CopyFile(path, treeviewControl.treeView.SelectedNode.Name + @"\" + System.IO.Path.GetFileName(path), Microsoft.VisualBasic.FileIO.UIOption.AllDialogs);
                                     }
-                                    catch
+                                    catch (System.Exception exception)
                                     {
-
+                                        System.Windows.Forms.MessageBox.Show(exception.Message, "List");
                                     }
                                 }
                             }
@@ -60,38 +89,34 @@
                         {
                             try
                             {
-                                System.IO.Directory.Delete(treeviewControl.treeView.SelectedNode.Name, true);
+                                System.Windows.Forms.DialogResult result = System.Windows.Forms.MessageBox.Show("'" + treeviewControl.treeView.SelectedNode.Text + "' will be deleted permanently.", "List", System.Windows.Forms.MessageBoxButtons.OKCancel, System.Windows.Forms.MessageBoxIcon.Warning);
+                                if (result == System.Windows.Forms.DialogResult.OK)
+                                {
+                                    System.IO.Directory.Delete(treeviewControl.treeView.SelectedNode.Name, true);
+                                }
                             }
-                            catch
+                            catch (System.Exception exception)
                             {
-
+                                System.Windows.Forms.MessageBox.Show(exception.Message, "List");
                             }
                         }
                     }
 
                     static private class File
                     {
-                        static public void Open(object sender, System.EventArgs e)
-                        {
-                            try
-                            {
-                                System.Diagnostics.Process.Start(treeviewControl.treeView.SelectedNode.Name);
-                            }
-                            catch
-                            {
-
-                            }
-                        }
-
                         static public void Delete(object sender, System.EventArgs e)
                         {
                             try
                             {
-                                System.IO.File.Delete(treeviewControl.treeView.SelectedNode.Name);
+                                System.Windows.Forms.DialogResult result = System.Windows.Forms.MessageBox.Show("'" + treeviewControl.treeView.SelectedNode.Text + "' will be deleted permanently.", "List", System.Windows.Forms.MessageBoxButtons.OKCancel);
+                                if (result == System.Windows.Forms.DialogResult.OK)
+                                {
+                                    System.IO.File.Delete(treeviewControl.treeView.SelectedNode.Name);
+                                }
                             }
-                            catch
+                            catch (System.Exception exception)
                             {
-
+                                System.Windows.Forms.MessageBox.Show(exception.Message, "List");
                             }
                         }
                     }
